% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ca_dsl.R
\name{read_ca_dsl}
\alias{read_ca_dsl}
\title{Read-In and Clean the Canada Domestic Substance List}
\usage{
read_ca_dsl(path, clean_non_ascii = FALSE)
}
\arguments{
\item{path}{(Character) The path to the .xlsx file.}

\item{clean_non_ascii}{(Logical) Should the non-ASCII characters be
reasonably converted? Defaults to \code{FALSE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the Canada
  Domestic Substance List.
}
\details{
The .xlsx file has to be manually downloaded from the
  following website:

  * https://pollution-waste.canada.ca/substances-search/Substance?lang=en
}
\examples{
\dontrun{
path <- "Domestic Substances List (DSL).xlsx"

dsl <- read_ca_dsl(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
