% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clp.R
\name{read_clp}
\alias{read_clp}
\title{Read-In and Clean the ECHA CLP Annex VI Data Set}
\usage{
read_clp(path, atp = TRUE)
}
\arguments{
\item{path}{(Character) The path to the XLSX file.}

\item{atp}{(Logical) Should the "atp" version information be included?
Defaults to \code{TRUE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the ECHA CLP
  Annex VI data set.
}
\details{
The function reads-in and cleans the ECHA CLP Annex VI data set
  into long flat format.
}
\note{
Tested with "atp" versions 9, 10, 13, 14, 15, and 17.
}
\examples{
\dontrun{
download.file(
  url = paste0(
    "https://echa.europa.eu/documents/10162/17218/",
    "annex_vi_clp_table_atp17_en.xlsx/",
    "4dcec79c-f277-ed68-5e1b-d435900dbe34?t=1638888918944"
  ),
  destfile = "annex_vi_clp_table_atp17_en.xlsx"
)

path <- "annex_vi_clp_table_atp17_en.xlsx"

clp <- read_clp(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
