% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cn_iecsc.R
\name{read_cn_iecsc}
\alias{read_cn_iecsc}
\title{Read-In and Clean the Chinese Inventory of Existing Chemical
  Substances and Chemicals}
\usage{
read_cn_iecsc(path)
}
\arguments{
\item{path}{(Character) The path to the PDF file.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the Chinese
  Inventory of Existing Chemical Substances and Chemicals data set.
}
\details{
The function reads-in and cleans the Chinese Inventory of Existing
  Chemical Substances and Chemicals data set into long flat format.
}
\note{
Tested with the 2013 PDF.
}
\examples{
\dontrun{
path <- "2013.pdf"

cn_iecsc <- read_cn_iecsc(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
