% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eu_eci.R
\name{read_eu_eci}
\alias{read_eu_eci}
\title{Read-In and Clean the ECHA EC Inventory Data Set}
\usage{
read_eu_eci(path, clean_non_ascii = FALSE)
}
\arguments{
\item{path}{(Character) The path to the CSV file.}

\item{clean_non_ascii}{(Logical) Should the non-ASCII characters be
reasonably converted? Defaults to \code{FALSE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the ECHA EC
  data set.
}
\details{
The function reads-in and cleans the ECHA EC data set into long
  flat format.
}
\note{
Tested with the "version" of 2022-01-11.
}
\examples{
\dontrun{
download.file(
  url = paste0(
    "https://echa.europa.eu/documents/10162/17222/ec_inventory_en.csv/",
    "326d9adb-27ed-5460-a2da-4f651b81e4b3"
    ),
  destfile = "ec_inventory_en.csv"
)

path <- "ec_inventory_en.csv"

eci <- read_eu_eci(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
