% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ioc.R
\name{read_ioc}
\alias{read_ioc}
\title{Read-In and Clean the New Zealand Inventory of Chemicals}
\usage{
read_ioc(path, clean_non_ascii = FALSE, version = TRUE)
}
\arguments{
\item{path}{(Character) The path to the XLSX file.}

\item{clean_non_ascii}{(Logical) Should the non-ASCII characters be
reasonably converted? Defaults to \code{FALSE}.}

\item{version}{(Logical) Should the version information be included?
Defaults to \code{TRUE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the New Zealand
  Inventory of Chemicals.
}
\details{
This function reads-in and automatically cleans the New Zealand
  Inventory of Chemicals.
}
\note{
Tested with the December 2021 version.
}
\examples{
\dontrun{
download.file(
  url = paste(
    "https://www.epa.govt.nz/assets/Uploads/Documents/Hazardous-Substances",
    "Guidance/NZIOC_Full_Spreadsheet_December_2021.xlsx",
    sep = "/"
  )
  destfile = "NZIOC_Full_Spreadsheet_December_2021.xlsx"
)

path <- "NZIOC_Full_Spreadsheet_December_2021.xlsx"

ioc <- read_ioc(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
