% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jp_cscl.R
\name{read_jp_cscl}
\alias{read_jp_cscl}
\title{Read-In and Clean the Japanese Chemical Substance Control Law
  Inventory}
\usage{
read_jp_cscl(path, clean_non_ascii = FALSE)
}
\arguments{
\item{path}{(Character) The path to the TSV file.}

\item{clean_non_ascii}{(Logical) Should the non-ASCII characters be
reasonably converted? Defaults to \code{FALSE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the Japanese
  Chemical Substance Control Law inventory.
}
\details{
The three TSV files have to be manually downloaded from the
  following website:

  * https://www.nite.go.jp/en/chem/chrip/chrip_search/sltLst
}
\examples{
\dontrun{
path <- "Japan CSCL_ Existing Chemical Substances.tsv"

cscl <- read_cscl(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
