% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jp_nite.R
\name{read_jp_nite}
\alias{read_jp_nite}
\title{Read-In and Clean the Japanese NITE Chemical Management GHS
  Classification Results}
\usage{
read_jp_nite(path, clean_non_ascii = FALSE)
}
\arguments{
\item{path}{(Character) The path to the XLSX file.}

\item{clean_non_ascii}{(Logical) Should the non-ASCII characters be
reasonably converted? Defaults to \code{FALSE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the Japanese
  NITE Chemical Management GHS Classification results.
}
\details{
This function reads-in and automatically cleans the Japanese NITE
  Chemical Management GHS Classification results.
}
\note{
Tested with the March 2022 version; GHS classifications omitted.
}
\examples{
\dontrun{
download.file(
  url = "https://www.nite.go.jp/chem/english/ghs/files/list_all_e.xlsx",
  destfile = "list_all_e.xlsx"
)

path <- "list_all_e"

nite <- read_jp_nite(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
