% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kr_ncis.R
\name{read_kr_ncis}
\alias{read_kr_ncis}
\title{Read-In and Clean the South Korean National Chemical Information
  System Data Set}
\usage{
read_kr_ncis(path, clean_non_ascii = FALSE)
}
\arguments{
\item{path}{(Character) The path to the XLSX file.}

\item{clean_non_ascii}{(Logical) Should the non-ASCII characters be
reasonably converted? Defaults to \code{FALSE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the South
  Korean National Chemical Information System (NCIS) data set.
}
\details{
The function reads-in and cleans the South Korean National Chemical
  Information System (NCIS) data set into long flat format.
}
\note{
Tested with an XLSX retrieved on 3 May 2022.
}
\examples{
\dontrun{
path <- "Chemical+Search_20220503055738.xlsx"

ncis <- read_kr_ncis(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
