% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nite.R
\name{read_nite}
\alias{read_nite}
\title{Read-In and Clean the Japanese NITE Chemical Management GHS
  Classification Results}
\usage{
read_nite(path, version = NULL)
}
\arguments{
\item{path}{(Character) The path to the XLSX file.}

\item{version}{(Character) Should the "atp" version information be included?
Defaults to \code{NULL}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the Japanese
  NITE Chemical Management GHS Classification results.
}
\details{
This function reads-in and automatically cleans the Japanese NITE
  Chemical Management GHS Classification results.
}
\note{
Tested with the March 2022 version; GHS classifications ommitted.
}
\examples{
\dontrun{
download.file(
  url = "https://www.nite.go.jp/chem/english/ghs/files/list_all_e.xlsx",
  destfile = "list_all_e.xlsx"
)

path <- "list_all_e"

nite <- read_nite(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
