% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ph_iccs.R
\name{read_ph_iccs}
\alias{read_ph_iccs}
\title{Read-In and Clean the Philippines Inventory of Chemicals and Chemical
  Substances}
\usage{
read_ph_iccs(path, file = "iccs", clean_non_ascii = FALSE, year = FALSE)
}
\arguments{
\item{path}{(Character) The path to the \code{iccs} PDF file. Not used
for \code{emb20} or \code{emb21}.}

\item{file}{(Character) Which type of file? Must be one of \code{piccs},
\code{emb20} or \code{emb21}.}

\item{clean_non_ascii}{(Logical) Should the non-ASCII characters be
reasonably converted? Defaults to \code{FALSE}.}

\item{year}{(Logical) Adds the year as extra column. Defaults to \code{FALSE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the Philippines
  Inventory of Chemicals and Chemical Substances.
}
\details{
This function reads-in and automatically cleans the Philippines
  Inventory of Chemicals and Chemical Substances.
}
\note{
Tested with the 2017 version.
}
\examples{
\dontrun{
download.file(
  url =
    "https://chemical.emb.gov.ph/wp-content/uploads/2020/07/PICCS.2017.pdf",
  destfile = "PICCS.2017.pdf"
)

path <- "PICCS.2017.pdf"

piccs <- read_piccs(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
