% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tsca.R
\name{read_tsca}
\alias{read_tsca}
\title{Read-In and Clean the US EPA TSCA Data Set}
\usage{
read_tsca(path, last_created = TRUE)
}
\arguments{
\item{path}{(Character) The path to the extracted CSV file.}

\item{last_created}{(Logical) Should the "last created" information be
extracted from the file name? Defaults to \code{TRUE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the US EPA
  TSCA data set.
}
\details{
The function reads-in and cleans the US EPA TSCA data set into flat
  format.
}
\note{
Tested with the "last created" version 2021-08.
}
\examples{
\dontrun{
zip_file <- download.file(
  url = paste0(
    "https://www.epa.gov/system/files/other-files/2021-08/",
    "csv-non-cbi-tsca-inventory-202108.zip"
    ),
  destfile = "csv-non-cbi-tsca-inventory-202108.zip"
)

unzip(zipfile = zip_file)

file_name <- list.files()

path <- file_name[grepl(pattern = "TSCA", file_name)]

tsca <- read_tsca(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
