% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tw_csi.R
\name{read_tw_csi}
\alias{read_tw_csi}
\title{Read-In and Clean the Taiwan Chemical Substance Inventory}
\usage{
read_tw_csi(
  path,
  pages = NULL,
  dpi = 600L,
  radius = 2L,
  threshold = 77L,
  whitelist = "- 0123456789CENP"
)
}
\arguments{
\item{path}{(Character) The path to the PDF file.}

\item{pages}{(Integer) Range of pages to read-in. See
\link[pdftools]{pdf_convert} for details. Defaults to \code{NULL}, i.e.,
all pages.}

\item{dpi}{(Integer) Resolution of the PDF image PNG render. See
\link[pdftools]{pdf_convert} for details. Defaults to \code{600L}.}

\item{radius}{(Integer) Noise reduction radius. See
\link[magick]{image_reducenoise} for details. Defaults to \code{2L}.}

\item{threshold}{(Integer) Black-and-white conversion threshold. See
\link[magick]{image_threshold} for details. Defaults to \code{77L}.}

\item{whitelist}{(Character) Limited allowed character set, optimized to
recognize CAS Registry Numbers. See \link[tesseract]{tesseract} for
details. Defaults to \code{"- 0123456789CENP"}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the Taiwan
  Chemical Substance Inventory.
}
\details{
This function reads-in and automatically cleans the Taiwan
  Chemical Substance Inventory.
}
\note{
Tested with the 1 January and 8 September 2019 versions.
}
\examples{
\dontrun{
download.file(
  url = paste(
    "https://gazette2.nat.gov.tw/EG_FileManager/eguploadpub/eg021170/ch08",
    "type3/gov82/num29/images/Eg01.pdf",
    sep = "/"
  ),
  destfile = "Eg01.pdf"
)tesseract

path <- "Eg01.pdf"

tcsi <- read_tw_csi(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
