% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_us_cdr.R
\name{read_us_cdr}
\alias{read_us_cdr}
\title{Read-In and Clean the United States Chemical Data Reporting Inventory}
\usage{
read_us_cdr(path)
}
\arguments{
\item{path}{(Character) The path to the CSV file.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the United
  States Chemical Data Reporting inventory.
}
\details{
The function reads-in and cleans the United States Chemical Data
  Reporting inventory.
}
\examples{
\dontrun{
path <- "2020 CDR Consumer and Commercial Use.csv"

cdr <- read_us_cdr(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
