% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xn_spin.R
\name{read_xn_spin}
\alias{read_xn_spin}
\title{Read-In and Clean the Substances in Preparations in Nordic Countries
  inventory}
\usage{
read_xn_spin(path, clean_non_ascii = FALSE)
}
\arguments{
\item{path}{(Character) The path to the TSV file.}

\item{clean_non_ascii}{(Logical) Should the non-ASCII characters be
reasonably converted? Defaults to \code{FALSE}.}
}
\value{
Returns a data frame.
}
\description{
This function reads-in and automatically cleans the Substances
  in Preparations in Nordic Countries inventory.
}
\details{
This function reads-in and automatically cleans the Substances in
  Preparations in Nordic Countries inventory.

  The file needs to be manually downloaded from the following URL:
  http://www.spin2000.net/spinmyphp/
}
\examples{
\dontrun{
path <- "spin.tsv"

spin <- read_xn_spin(path)
}
}
\author{
Raoul Wolf (\url{https://github.com/RaoulWolf/})
}
