% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsca.R
\docType{data}
\name{tsca}
\alias{tsca}
\title{The US EPA TSAC Data Set}
\format{
A data frame with 68,191 rows and 11 variables:
\describe{
  \item{id}{record ID number}
  \item{casrn}{Chemical Abstracts Service (CAS) registry number}
  \item{casregno}{CAS registry number without "-" (dashes)}
  \item{uid}{unique identifier}
  \item{exp}{expiration date}
  \item{chemname}{preferred Chemical Abstracts (CA) index name}
  \item{def}{chemical substance definition}
  \item{uvcb}{UVCB flag}
  \item{flag}{EPA TSCA regulatory flag}
  \item{activity}{commercial activity status}
  \item{last_created}{year and month (YYYY-MM) of data set publication}
}
}
\source{
\url{https://www.epa.gov/tsca-inventory}
}
\usage{
tsca
}
\description{
The Toxic Substances Control Act (TSCA) Chemical Substance
  Inventory contains all existing chemical substances manufactured, processed,
  or imported in the United States that do not qualify for an exemption or
  exclusion under TSCA.

This data set includes the non-confidential portion of the inventory.

More information on the columns can be found here:
  \url{https://www.epa.gov/tsca-inventory/how-access-tsca-inventory}.
}
\keyword{datasets}
